requirements.txt
flet
pandas
matplotlib
Pillow
numpy